Release 1 | Beta 0 | Development 2 | Experimental 3
* Changed getDelta() methods to be of type float to be
  more accurate

* Fixed bug: Terrain normals were incorrectly calculated
* Fixed bug: DebugCamera was incorrectly calculating mouse
             movement
* Fixed bug: GUIDropDownMenu and hence GUIDropDownList did
             not assign added button's transform
* Fixed bug: Horizontal GUISlider did not work correctly

+++ UnnamedEngine/Source/core/BaseEngine.cpp
@@ -150,7 +150,7 @@ void BaseEngine::renderDebuggingInfo() {
 							"Engine Version : " + str(Engine::Version) + "\n" +
 							"Engine Date    : " + str(Engine::DateCreated) + "\n" +
 							"Engine Build   : " + str(Engine::Build) + "\n" +
-							"Current Delta  : " + str(getDelta()) + "\n" +
+							"Current Delta  : " + str((int) getDelta()) + "\n" +
 							"Current FPS    : " + str(getFPS()) + "\n" +
 							"----------- VIDEO -----------\n" +
 							"Resolution     : " + str(getSettings().videoResolution.getX()) + "x" + str(getSettings().videoResolution.getY()) + "\n" +
